project_m_n_e_service/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-environ==0.9.0
django-ninja==0.18.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pydantic==1.9.1
sqlparse==0.4.2
typing_extensions==4.2.0
